requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.1.2
django-allauth==0.51.0
django-countries==7.4.2
django-filter==22.1
django-recaptcha==3.0.0
djangorestframework==3.14.0
idna==3.4
Markdown==3.4.1
oauthlib==3.2.2
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.5
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.12
